requirements-37.txt
numpy
pandas
matplotlib
scikit-learn
librosa
python_speech_features
pyautogui
pydirectinput
dragonfly2
joblib